Add files via upload
Coding Project
Write a console app that models the system on the next slide.
You should first initialise your system passing in the depth of the system – in this case 4. 
Set the state of each gate switch randomly to left or right.
Model 2 behaviours:
Predict the answer to the solution i.e which container does not receive a ball.
Run the model to completion and check that your prediction matches the
outcome.
At each intersection is a gate.
When a ball passes a gate it will pass down the open side, but then the gate will move to the other side.
There are 15 balls and 16 containers. Which container will not receive a ball when all balls have passed through the system.